+++ spec/components/atoms/TreeNode.spec.ts
@@ -18,7 +18,9 @@ Copyright (C) 2021, Tomoya Komiyama.
 */
 
 import { mount } from '@vue/test-utils'
+import { ref } from 'vue'
 import Target from '/@/components/atoms/TreeNode.vue'
+import { provideTreeContext } from '/@/composables/treeContext'
 
 describe('src/components/atoms/TreeNode.vue', () => {
   describe('snapshot', () => {
@@ -38,6 +40,7 @@ describe('src/components/atoms/TreeNode.vue', () => {
       expect(wrapper.element).toMatchSnapshot()
     })
     it('closed nested tree', () => {
+      const closed = ref<{ [id: string]: boolean }>({})
       const wrapper = mount<any>(Target, {
         props: {
           node: {
@@ -47,6 +50,12 @@ describe('src/components/atoms/TreeNode.vue', () => {
               { id: 'aa', name: 'name_aa', children: [] },
               { id: 'bb', name: 'name_bb', children: [] },
             ],
+          },
+        },
+        global: {
+          provide: {
+            getClosedMap: () => closed.value,
+            setClosed: (id: string, val: boolean) => (closed.value[id] = val),
           },
         },
       })
@@ -84,6 +93,55 @@ describe('src/components/atoms/TreeNode.vue', () => {
       await wrapper.find('.node-name').trigger('dblclick')
       expect(wrapper.find('.node-name').exists()).toBe(true)
       expect(wrapper.find('input').exists()).toBe(false)
+    })
+  })
+
+  describe('expand all', () => {
+    const node = {
+      id: 'a',
+      name: 'name_a',
+      children: [
+        {
+          id: 'aa',
+          name: 'name_aa',
+          children: [
+            {
+              id: 'aaa',
+              name: 'name_aaa',
+              children: [{ id: 'aaaa', name: 'name_aaaa', children: [] }],
+            },
+          ],
+        },
+        { id: 'bb', name: 'name_bb', children: [] },
+      ],
+    }
+
+    it('should expand all along newly selected node', async () => {
+      const selected = ref<{ [id: string]: boolean }>({})
+      const provided: { [key: string]: any } = {}
+      provideTreeContext('test', {
+        getSelectedMap: () => selected.value,
+        provide: (key, value) => (provided[key] = value),
+      })
+      provided.setClosed('a', true)
+      provided.setClosed('aa', true)
+      provided.setClosed('aaa', true)
+      provided.setClosed('aaaa', true)
+      provided.setClosed('b', true)
+      provided.setClosed('bb', true)
+
+      const wrapper = mount(Target, {
+        props: { node },
+        global: { provide: provided },
+      })
+      selected.value['aaa'] = true
+      await wrapper.vm.$nextTick()
+      expect(provided.getClosedMap()).toEqual({
+        aaa: true,
+        aaaa: true,
+        b: true,
+        bb: true,
+      })
     })
   })
 })